First Installation
Coinciding with first basic success of MarketSimulator. Have resolved
issue that was occurring with cleaning Mock Methods, by overwriting
generic S3 method with default, and then detaching the mock methods
environment. Also created a register of generics which are created in 
order to later be able to remove them.
Documentation has been cleaned up mostly as a result of running 
package checks. There are still some niggles which should be sorted:
1. Maybe mocks should default to standard function rather than throwing
an error if an unregistered mock method is called on them.
2. Checking call registers requires that arguments are named in the 
test if they are named in the code. Should be independent of this.